DBZ-677 Extracting shared configuration options for both new state extraction SMTs
